good stuff

+++ app/controllers/photos_controller.rb
@@ -16,13 +16,26 @@ class PhotosController < ApplicationController
   end
 
   def destroy
+    redirect_to '/gallery' and return if Photo.count == 1
     photo = Photo.find(params[:id])
-    photo.photo_placers.update_all(photo_id: Photo.first.id)
+    update_placeholders(photo)
     photo.destroy
     redirect_to '/gallery'
   end
 
   private
+
+  def update_placeholders(photo)
+    photo.photo_placers.update_all(photo_id: safe_photo(photo).id)
+  end
+
+  def safe_photo(photo)
+    photo == first_photo ?  Photo.last : first_photo
+  end
+
+  def first_photo
+    @first_photo ||= Photo.first
+  end
 
   def photo_params
     params.require(:photo).permit(:image, :caption)
